BackendofADEIVotechain/src/main/java/com/Votechainbackend/BackendofADEIVotechain/services/LoginService.java
package com.Votechainbackend.BackendofADEIVotechain.services;

import com.Votechainbackend.BackendofADEIVotechain.entities.User;
import com.Votechainbackend.BackendofADEIVotechain.repositories.UserRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.Optional;

@Service
public class LoginService {
    private final UserRepository userRepository;

    @Autowired
    public LoginService(UserRepository userRepository) {
        this.userRepository = userRepository;
    }

    public User verifyLogin(String email, String password) {
        Optional<User> userOptional = userRepository.findByEmailAddress(email);
        if(userOptional.isPresent()){
            User user = userOptional.get();
            if(user.getPassword().equals(password)){
                return user;
            } else {
                throw new IllegalStateException("Invalid password!");
            }
        } else {
            throw new IllegalStateException("User not found!");
        }
    }
}

